+++ src/main/java/dinkplugin/notifiers/LeaguesNotifier.java
@@ -331,11 +331,13 @@ public class LeaguesNotifier extends BaseNotifier {
             var tierStruct = client.getStructComposition(tierStructs[tierIndex - 1]);
             pointsMap.put(tierStruct.getIntValue(877), tier);
 
+            /*
             var relicStructs = client.getEnum(tierStruct.getIntValue(878)).getIntVals();
             for (int relicStruct : relicStructs) {
                 var name = client.getStructComposition(relicStruct).getStringValue(879);
                 TIER_BY_RELIC.put(name, tier);
             }
+            */
         }
         if (pointsMap.size() > 1) {
             TIER_BY_POINTS.clear();
@@ -362,37 +364,29 @@ public class LeaguesNotifier extends BaseNotifier {
             .collect(Collectors.toMap(LeagueRelicTier::getDefaultPoints, Function.identity(), (a, b) -> null, TreeMap::new));
 
         TIER_BY_RELIC = new HashMap<>(Map.ofEntries(
-            // confirmed tier 1 relics
             Map.entry("Animal Wrangler", LeagueRelicTier.ONE),
             Map.entry("Lumberjack", LeagueRelicTier.ONE),
             Map.entry("Power Miner", LeagueRelicTier.ONE),
-
-            // confirmed tier 2 relics
             Map.entry("Corner Cutter", LeagueRelicTier.TWO),
             Map.entry("Dodgy Deals", LeagueRelicTier.TWO),
             Map.entry("Friendly Forager", LeagueRelicTier.TWO),
-
-            // confirmed tier 3 relics
             Map.entry("Bank Heist", LeagueRelicTier.THREE),
             Map.entry("Clue Compass", LeagueRelicTier.THREE),
             Map.entry("Fairy's Flight", LeagueRelicTier.THREE),
-
-            // confirmed tier 4 relics; missing 1 relic
+            Map.entry("Equilibrium", LeagueRelicTier.FOUR),
             Map.entry("Golden God", LeagueRelicTier.FOUR),
             Map.entry("Reloaded", LeagueRelicTier.FOUR),
-
-            // total recall vs banker's note tier; probably tier 5
-            Map.entry("Total Recall", LeagueRelicTier.UNKNOWN),
-            Map.entry("Banker's Note", LeagueRelicTier.UNKNOWN),
-
-            // grimiore vs overgrown vs ??? tier; probably tier 6
-            Map.entry("Grimoire", LeagueRelicTier.UNKNOWN),
-            Map.entry("Overgrown", LeagueRelicTier.UNKNOWN),
-
-            // combat tier; probably tier 7
-            Map.entry("Guardian", LeagueRelicTier.UNKNOWN),
-            Map.entry("Last Stand", LeagueRelicTier.UNKNOWN),
-            Map.entry("Specialist", LeagueRelicTier.UNKNOWN)
+            Map.entry("Production Master", LeagueRelicTier.FIVE),
+            Map.entry("Slayer Master", LeagueRelicTier.FIVE),
+            Map.entry("Treasure Arbiter", LeagueRelicTier.FIVE),
+            Map.entry("Total Recall", LeagueRelicTier.SIX),
+            Map.entry("Banker's Note", LeagueRelicTier.SIX),
+            Map.entry("Grimoire", LeagueRelicTier.SEVEN),
+            Map.entry("Overgrown", LeagueRelicTier.SEVEN),
+            Map.entry("Pocket Kingdom", LeagueRelicTier.SEVEN),
+            Map.entry("Guardian", LeagueRelicTier.EIGHT),
+            Map.entry("Last Stand", LeagueRelicTier.EIGHT),
+            Map.entry("Specialist", LeagueRelicTier.EIGHT)
         ));
     }
 }